Addresses must allow for long.
They're 32-bit *unsigned* integers, which don't exist in Java. Mostly
this doesn't matter, but when JSON transfers a value it matters a lot.
The transceiver API should just cope.

+++ SpiNNaker-front-end/src/main/java/uk/ac/manchester/spinnaker/front_end/download/RecordingRegionDataGatherer.java
@@ -89,7 +89,7 @@ public class RecordingRegionDataGatherer extends DataGatherer
 	}
 
 	private synchronized RecordingRegionsDescriptor getDescriptor(
-			ChipLocation chip, int baseAddress)
+			ChipLocation chip, long baseAddress)
 			throws IOException, ProcessException {
 		RRKey key = new RRKey(chip, baseAddress);
 		RecordingRegionsDescriptor rrd = descriptors.get(key);
@@ -218,9 +218,9 @@ public class RecordingRegionDataGatherer extends DataGatherer
 	 */
 	private static final class RRKey {
 		private final ChipLocation chip;
-		private final int baseAddr;
+		private final long baseAddr;
 
-		RRKey(HasChipLocation chip, int baseAddress) {
+		RRKey(HasChipLocation chip, long baseAddress) {
 			this.chip = chip.asChipLocation();
 			this.baseAddr = baseAddress;
 		}
@@ -236,7 +236,7 @@ public class RecordingRegionDataGatherer extends DataGatherer
 
 		@Override
 		public int hashCode() {
-			return baseAddr ^ chip.hashCode();
+			return ((int) baseAddr) ^ chip.hashCode();
 		}
 	}
 }
